# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Volleyball Tryout Manager API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /sync/trigger:
    post:
      operationId: triggerSync
      tags: [sync]
      summary: Trigger manual data sync
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SyncTrigger"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncStatus"
  /sync/sheets:
    post:
      operationId: syncFromGoogleSheets
      tags: [sync]
      summary: Sync players from a Google Sheet
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SheetsSyncRequest"
      responses:
        "200":
          description: Sync result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportResult"

  # AI
...
components:
  schemas:
...
    ImportResult:
      type: object
      required: [imported, updated, errors]
      properties:
        imported:
          type: integer
        updated:
          type: integer
        errors:
          type: array
          items:
            type: string
...
    SyncStatus:
      type: object
      required: [lastSyncAt, status, playersUpdated]
      properties:
        lastSyncAt:
          type: ["string", "null"]
          format: date-time
        status:
          type: string
          enum: [idle, syncing, success, error]
        playersUpdated:
          type: integer
        message:
          type: ["string", "null"]
    SyncTrigger:
      type: object
      properties:
        source:
          type: string
          enum: [manual, auto]
    SheetsSyncRequest:
      type: object
      required: [sheetUrl]
      properties:
        sheetUrl:
          type: string
          description: Google Sheet URL or spreadsheet ID
        sheetName:
          type: string
          description: Sheet tab name (defaults to first sheet)
...